chore: move TeamForm from teams/create/_component to teams/_component
