+++ app/(general)/performances/[id]/teams/_components/TeamForm/FirstPage/schema.ts
