+++ app/(general)/performances/[id]/teams/_components/TeamForm/SecondPage/MemberSessionRequiredCheckbox.tsx
