+++ app/(general)/performances/[id]/teams/_components/TeamForm/SecondPage/schema.ts
