+++ app/(general)/performances/[id]/teams/create/page.tsx
@@ -6,7 +6,7 @@ import { auth } from "@/auth"
 import OleoPageHeader from "@/components/OleoPageHeader"
 import ROUTES from "@/constants/routes"
 
-import TeamForm from "./_components/TeamForm"
+import TeamForm from "../_components/TeamForm"
 
 interface TeamCreatePageProps {
   params: {
